Added maps navigation
Selecting a pokemon now brings you to a navigation screen with a route to the Pokemon using Google Maps Directions API. Just follow the line. The black dot in the middle is your location, the circle around you is the 50m radius in which Pokemon are visible. Using the up and down keys when tracking will zoom the map in and out.

WARNING: It appears that the app is suffering from this issue:
https://github.com/pebble/pebblejs/issues/161
So the app will eventually crash after returning from the Maps window to the Pokemon menu.